include module exports in instructions
in issues https://github.com/askmike/gekko/issues/2493 and https://github.com/askmike/gekko/issues/2576, users saw an error (`TypeError: Cannot read property 'timeout' of undefined`) that is not easy to understand because they had edited their uiconfig files according to the instructions here, which omit the export. This change is to include an example of a complete uiconfig file to reduce confusion by users who copy and paste the example code over the entire contents of their existing uiconfig.